feat: construct ABC Species
Species is an abstract base class. PreySpecies and Predator Species
are subclasses of Species. The Species is implemented to make
comparisons and calculations between instances of Species or
int/float.

Note that currently, LampreySpecies is NOT a subclass of Species as the
born/spawn stage of Lamprey is much more complex than its prey and
predator.