feat: implement Rule, Rulesets class
The Rulesets Class is a list of instances of Rule. Every Rule contains
two callback function, condition and action. The action is applied when
the condition is met.

Since almost all our rules' conditions are based on month, we implement
a function: which_month() to judge whether the current month is
satisfiable.

BREAKING CHANGE: change the step() in Ecosystem to use Rulesets