feat: implement Metrics Class
An instance of World should have an instance of Metrics as its attribute
to keep track of basic info and metrics as the world iterates.

This version of Metrics Class is very simple and may change in the next
version. As it currently does not need parameters at init, the metric of
a world is created inside the init of World rather than a parameter.